+++ src/__init__.py
@@ -1,0 +1,10 @@
+"""
+src
+Samuel Koller
+Created: 17 October 2024
+Updated: 17 October 2024
+
+Contains all source code used for simulating a running elevator.
+"""
+
+from .classes import *